Four new lightbox examples have been added
1) lightbox with focus areas within it, but when the lightbox opens focus
remains in the background page area requiring the user to tab through a
number of links before getting to the focus areas within the light box.

2) A lightbox where the X doesn't receive keyboard focus but it can be closed
by clicking on it with a mouse or touching using a touch screen.

3) Focus moves to a focus area in the lightbox, but focus is not retained
within the lightbox, ie tabbing will take the user to other focus areas on the
page

4) lightbox can't be closed using the escape key (this may not be a
requirement but it is a commonly known way of closing light boxes)